Consolidate blocks into ee/handle_standard_answers
